refactor(mcmc): rename hmc.{h,cpp} to hamiltonian_utils.{h,cpp}
The file holds kinetic_energy() and the heuristic_initial_step_size()
overloads - Hamiltonian utilities consumed by the NUTS code; there is
no HMC sampler. Includes updated in nuts.cpp, nuts_sampler.h,
omrf_model.cpp, and bgmCompare_sampler.cpp; sources.mk regenerated
via inst/generate_makevars_sources.R.
